Wrote a script to generate a multi-AS topology running OSPF/BGP. <NOT DEBUGGED YET!>
This is the very first version of the multi-AS topology script. There are a few errors with the TCL/ns-2 syntax that are not yet fixed. The reason this version of the script is committed is because the syntax organisation is about to change somewhat fundamentally, hence this will probably serve as the best "revert-point."