Changed the network topology script to create meaningful, IP-space-preserving links, and improved customisability.
The script can now support the creation of a few hundred nodes and/or ASes, easily customisable through the script's constant parameters. (The topology does adhere to a few conventions, however. These are noted in the script's description.) Part of the script's functionality is the exporting of information about all links between nodes as well as their respective IPs.

**PROBLEM** The 'export_array' procedure does not work on Emulab, although it does on local hosts that I tested it on. It seems, also that this is not due to the permissions that are set for the specified output log file. Interestingly (and annoingly) enough, the file is created, but ends up being empty.